Bypassing assertion when checking strings.
git-svn-id: https://svn.syssoft.uni-trier.de/svn/prisma/DeltaWorks/include@411 87af6b1e-9b34-e211-997a-bc305bce1b9a

+++ net/containers.h
@@ -11,7 +11,7 @@ namespace Package
 		protected:
 		static	bool			validChar(char c)
 				{
-					return isgraph(c) || isspace(c);
+					return isgraph((BYTE)c) || isspace((BYTE)c);
 				}
 		public:
 								NetString()
